PublisherCompras/requirements.txt
fastapi
uvicorn
psycopg2
paho-mqtt==1.6.1
requests==2.31.0
python-dotenv
